requirements.txt
streamlit>=1.36.0
transformers>=4.43.0
torch>=2.2.0
sentencepiece>=0.1.99
pdfplumber>=0.11.0
python-docx>=1.1.0
openai>=1.30.0
reportlab>=4.0.9
accelerate>=0.20.0








